Separated the initial survey models in 2, for coach and client
